% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{compute_C}
\alias{compute_C}
\title{compute geary's C for a given gene.}
\usage{
compute_C(feature = NULL, SpaCoObject)
}
\arguments{
\item{feature}{gene to compute geary's C for.}

\item{SpaCoObject}{containing the data.}
}
\value{
returns geary's c value.
}
\description{
compute geary's C for a given gene.
}
