% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{compute_nSpacs}
\alias{compute_nSpacs}
\title{compute_nSpacs}
\usage{
compute_nSpacs(
  SpacoObject,
  nSim,
  PC_criterion = "percent",
  PC_value = 0.8,
  SpacQuantile = 0.5
)
}
\arguments{
\item{nSim}{number of simulations i.e. random permutation matrices}

\item{PC_criterion}{criterion on which to select number of principal components for initial covariance matrix reconstruction; either "number" to select a number of PCs or "percent" to select number of PCs to explain specified amount of data variance}

\item{PC_value}{Value to specify number of PCs or desired level of explained variance, see "PC_criterion"}

\item{SpacQuantile}{quantile of simulated Spac distribution used for determining cutoff}
}
\description{
compute_nSpacs
}
