% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{compute_smoothed_profiles}
\alias{compute_smoothed_profiles}
\title{Title}
\usage{
compute_smoothed_profiles(SpaCoObject, nSpacs)
}
\arguments{
\item{SpaCoObject}{Spaco object to compute profiles of}

\item{nSpacs}{number of spac's to use to compute the smoothed profiles. Should be determined by the compute_nSpacs function.}
}
\description{
Title
}
