% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seuratwrappers.R
\name{create_SpaCoObject_from_KNN}
\alias{create_SpaCoObject_from_KNN}
\title{create_SpaCoObject_from_KNN}
\usage{
create_SpaCoObject_from_KNN(SeuratObject, n = 10)
}
\arguments{
\item{SeuratObject}{Seurat object to export kNN-graph from.}

\item{n}{Number of neighbors to consider.}
}
\value{
Returns a SPaCoObject with the SCT data and the kNN-graph as neighborhood matrix.
}
\description{
create_SpaCoObject_from_KNN
}
