% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{denoise_profiles}
\alias{denoise_profiles}
\title{computes smoothed gene profiles of genes present in the data.}
\usage{
denoise_profiles(SpaCoObject)
}
\arguments{
\item{SpaCoObject}{Spaco object to compute profiles of.}
}
\value{
smoothed gene profiles in the SpaCoObject.
}
\description{
computes smoothed gene profiles of genes present in the data.
}
