% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{denoised_projection_plot}
\alias{denoised_projection_plot}
\title{Plot denoised gene expression}
\usage{
denoised_projection_plot(SpaCoObject, feature = NULL)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{feature}{Gene for which to plot denoised projection}
}
\value{
returns a ggplot object with the denoised gene expression.
}
\description{
Plot denoised gene expression
}
