% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{.orthogonalizeA}
\alias{.orthogonalizeA}
\title{Title}
\usage{
.orthogonalizeA(X, A, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{X}{Projection to orthogonalize}

\item{A}{GraphLaplacian}

\item{tol}{tolerance}
}
\value{
An orthogonolized A matrix.
}
\description{
Title
}
\keyword{internal}
