% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Multiply two matrices using Eigen library}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\value{
the product of A and B
}
\description{
This function multiplies two matrices using the Eigen library, which provides
fast linear algebra operations.
}
