% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{feature_predict}
\alias{feature_predict}
\title{Prediction for IMPACT Dataset}
\usage{
feature_predict(spaco_object)
}
\arguments{
\item{spaco_object}{A list containing the results of model training. It is
assumed to include elements \code{num_subjects}, \code{num_times}, \code{K}, \code{X}, \code{V},
\code{Phi}, and \code{mu}.}
}
\value{
A list containing the predictive values.
}
\description{
This function evaluates the predictive power of SPACO model
trained on IMPACT dataset. The observation for each subject is predicted by
\deqn{\hat{x}_{itj} =  \sum_{k=1}^K \hat{u}_{ik} \phi_{kt} v_{jk}}
}
\examples{
# IMPACT
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
spaco_object <- train_prepare(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                              Z = impact$Z, K = 4, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
spaco_object <- feature_predict(spaco_object)

# IMMUNE
data("immune_original")
immune <- immune_data_wrangling(immune_original)
spaco_object <- train_prepare(X = immune$X, OBS = immune$OBS, T1 = immune$T1,
                              Z = immune$Z, K = 5, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
spaco_object <- feature_predict(spaco_object)
}
