% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalFunctions.R
\name{FindSVG}
\alias{FindSVG}
\title{wrapper function to compute spatial variable genes and their empirical p-values}
\usage{
FindSVG(SpaCoObject, nSpacs, nSim = 1000)
}
\arguments{
\item{nSim}{}
}
\description{
wrapper function to compute spatial variable genes and their empirical p-values
}
