% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune}
\alias{immune}
\title{immune}
\description{
immune
}
\details{
\code{time} is the time of the measurement;
\code{X} acceleration in the "X" direction;
\code{Y} acceleration in the "Y" direction;
\code{Z} acceleration in the "Z" direction;
}
\keyword{data}
