% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{immune_data_wrangling}
\alias{immune_data_wrangling}
\title{Data Wrangling for IMMUNE Dataset}
\usage{
immune_data_wrangling(immune_original)
}
\arguments{
\item{immune_original}{Original IMMUNE dataset.}
}
\value{
A list \code{immune} containing the following elements:
\item{X}{A 36 (number of subjects) × 40 (number of times) × 133 (number
of features) array containing time-series data.}
\item{OBS}{A 36 (number of subjects) × 40 (number of times) × 133 (number
of features) array indicating whether an observation is available in array
\code{X}.}
\item{T1}{A length 40 (number of times) vector containing measured time.}
\item{Z}{A 36 (number of subjects) × 10 (number of covariates) matrix
containing auxiliary covariates.}
\item{columns_feature}{Name of all features.}
}
\description{
This function transforms and maps raw data form of IMMUNE into
the desired format. Features with missing values are excluded. Several
covariates are selected including \code{age}, \code{sex}, \code{obesity}, \code{hospitalized},
\code{severity_group}, \code{intubated}, \code{alive}, \code{tocilizumab}, \code{heme}, and \code{bmt}.
}
\examples{
# IMMUNE
data("immune_original")
immune <- immune_data_wrangling(immune_original)
spaco_object <- train_prepare(X = immune$X, OBS = immune$OBS, T1 = immune$T1,
                              Z = immune$Z, K = 5, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
}
