% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_original}
\alias{immune_original}
\title{Original IMMUNE Dataset}
\description{
IMMUNE dataset retrieved from the original source.
}
\details{
See \href{https://pubmed.ncbi.nlm.nih.gov/33361110/}{IMMUNE Profiling Study}
for more information.
The dataset contains 106 samples of 170 immunological observations collected
from 45 participants infected with COVID-19. Blood samples of patients were
collected at enrollment and approximately every 7 days thereafter. The
ratios of different immune cells in the blood test were measured at these
time points.
}
\keyword{data}
