% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact_data_wrangling}
\alias{impact_data_wrangling}
\title{Data Wrangling for IMPACT Dataset}
\usage{
impact_data_wrangling(impact_missing, impact_imputed)
}
\arguments{
\item{impact_missing}{IMPACT dataset with missing values.}

\item{impact_imputed}{IMPACT dataset after imputing missing values.}
}
\value{
A list \code{impact} containing the following elements:
\item{X}{A 98 (number of subjects) × 35 (number of times) × 111 (number
of features) array containing time-series data.}
\item{OBS}{A 98 (number of subjects) × 35 (number of times) × 111 (number
of features) array indicating whether an observation is available in array
\code{X}.}
\item{T1}{A length 35 (number of times) vector containing measured time.}
\item{Z}{A 98 (number of subjects) × 8 (number of covariates) matrix
containing auxiliary covariates.}
\item{Y}{A 98 (number of subjects) × 4 (number of responses) matrix
containing responses.}
\item{imputed_pt}{Patient data obtained from IMPACT dataset.}
\item{filtered_feature_idx}{Indices of filtered features for model training.}
}
\description{
This function transforms and maps raw data form of IMPACT into
the desired format. Features with more than 20\% missingness are excluded.
Several covariates are selected including \code{COVIDRISK_1}, \code{COVIDRISK_2},
\code{COVIDRISK_3}, \code{COVIDRISK_4}, \code{COVIDRISK_5}, \code{Age}, \code{sex}, and \code{BMI}.
}
\examples{
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
spaco_object <- train_prepare(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                              Z = impact$Z, K = 4, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
}
