% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_imputed}
\alias{impact_imputed}
\title{Imputed IMPACT Dataset}
\description{
IMPACT dataset after imputing missing values.
}
\details{
See \href{https://pubmed.ncbi.nlm.nih.gov/32717743/}{IMPACT Study} for more
information.
The dataset consists of 180 samples across 135 immunological features
collected from 98 participants infected with COVID-19. Viral RNA load,
levels of plasma cytokines and chemokines, and leukocyte profiles were
assessed in the study. Follow-up measurements of the patients included 1 to
7 longitudinal time points, occurring 3-51 days after symptom onset.
}
\keyword{data}
