% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_missing}
\alias{impact_missing}
\title{impact_missing}
\description{
impact_missing
}
\details{
\code{time} is the time of the measurement; \cr
\code{X} acceleration in the "X" direction; \cr
\code{Y} acceleration in the "Y" direction; \cr
\code{Z} acceleration in the "Z" direction;
}
\keyword{data}
