% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{impact_plot}
\alias{impact_plot}
\title{Plot for IMPACT Dataset}
\usage{
impact_plot(spaco_object, feature_name, imputed_pt, filtered_feature_idx)
}
\arguments{
\item{spaco_object}{A list containing the results of model training. It is
assumed to include elements \code{num_subjects}, \code{num_times}, \code{X}, \code{OBS}, and
\code{Xhat}.}

\item{feature_name}{Name of the selected feature.}

\item{imputed_pt}{Patient data obtained from IMPACT dataset.}

\item{filtered_feature_idx}{Indices of filtered features for model training.}
}
\value{
A plot of the observed versus estimated values.
}
\description{
This function plots the observed versus estimated values for
the selected feature.
}
\examples{
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
spaco_object <- train_prepare(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                              Z = impact$Z, K = 4, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
spaco_object <- impact_predict(spaco_object)
impact_plot(spaco_object, "TcellsofLivecells",
            impact$imputed_pt, impact$filtered_feature_idx)
impact_plot(spaco_object, "TotalNeutrophilsofLivecells",
            impact$imputed_pt, impact$filtered_feature_idx)
impact_plot(spaco_object, "HLA.DR.ofTotalMono",
            impact$imputed_pt, impact$filtered_feature_idx)
impact_plot(spaco_object, "IL6",
            impact$imputed_pt, impact$filtered_feature_idx)
}
