% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedSpaCoObject_constructor.R
\name{mergeSpaCoObjects}
\alias{mergeSpaCoObjects}
\title{Create a constructor function that creates an object of class SpaCoObject}
\usage{
mergeSpaCoObjects(object1, object2, label1 = "dataset1", label2 = "dataset2")
}
\arguments{
\item{neighbours}{Binary matrix with weights describing if cells are to be considered neighbours or not depending on the defined distance.}

\item{data}{Matrix with normalized and scaled gene counts. Rows as cells and genes as columns}

\item{coordinates}{Matri with the cell coordinates on the slides. Rows and Columns in the 10x Visium case.}
}
\value{
Returns a SpaCoObject with the given slots filled
}
\description{
Create a constructor function that creates an object of class SpaCoObject
}
