% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{normA}
\alias{normA}
\title{Title}
\usage{
normA(x, A, preFactor)
}
\arguments{
\item{x}{placeholder}

\item{A}{placeholder}
}
\value{
placeholder
}
\description{
Title
}
\examples{
placeholder
}
\keyword{internal}
