% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{normalizeA}
\alias{normalizeA}
\title{normalizeA}
\usage{
normalizeA(x, A, preFactor)
}
\arguments{
\item{x}{placeholder}

\item{A}{placeholder}
}
\value{
placeholder
}
\description{
normalizeA
}
\examples{
placeholder
}
\keyword{internal}
