% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalwrapper.R
\name{PValWrapperFunction_object}
\alias{PValWrapperFunction_object}
\title{wrapper function to compute spatial variable genes and their empirical p-values}
\usage{
PValWrapperFunction_object(SpaCoObject, nSpacs, nSim = 1000)
}
\arguments{
\item{nSim}{}
}
\description{
wrapper function to compute spatial variable genes and their empirical p-values
}
