% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{rank_selection}
\alias{rank_selection}
\title{Rank Selection via Cross-validation}
\usage{
rank_selection(
  X,
  OBS,
  T1,
  Z,
  ranks,
  early_stop = TRUE,
  max_iter = 30,
  cv_iter = 5,
  nfolds = 5,
  extra_std = 0,
  random_state = 2022,
  trace = TRUE
)
}
\arguments{
\item{X}{An \code{I} (number of subjects) × \code{T} (number of times) × \code{J} (number
of features) array containing time-series data.}

\item{OBS}{An \code{I} (number of subjects) × \code{T} (number of times) × \code{J} (number
of features) array indicating whether an observation is available in array
\code{X}.}

\item{T1}{A length \code{T} (number of times) vector containing measured time.}

\item{Z}{An \code{I} (number of subjects) × \code{q} (number of covariates) matrix
containing auxiliary covariates.}

\item{ranks}{An array of ranks containing integers greater than 1 from
which to select potential ranks.}

\item{early_stop}{Whether to stop the process when the cross-validated
marginal log-likelihood is no longer decreasing. Default is TRUE.}

\item{max_iter}{Maximum number of iterations in the model training
process. Default is 30.}

\item{cv_iter}{Maximum number of iterations in the cross-validated
training process. Default is 5.}

\item{nfolds}{Number of folds. Default is 5.}

\item{extra_std}{Coefficient of standard deviation when calculating the
standard deviation of log-likelihood. Default is 0.}

\item{random_state}{Control the randomness of each fold. When shuffling the
data before splitting into batches, it affects the ordering of the indices.
Default is 2022.}

\item{trace}{Whether to print the model training process. Default is TRUE.}
}
\value{
An integer indicating selected rank.
}
\description{
This function performs rank selection as suggested in SupCP
(Lock and Li, 2018). Model parameters are estimated on the training sets
and evaluated on the test sets. The rank is chosen to be the value that
achieves the lowest log-likelihood of test sets.
}
\examples{
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
ranks <- c(2:10)
rank <- rank_selection(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                       Z = impact$Z, ranks = ranks, early_stop = TRUE)
}
