% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_for_spaco.R
\name{read_10x_for_spaco}
\alias{read_10x_for_spaco}
\title{Read in 10x Visium spatial transcriptomics data}
\usage{
read_10x_for_spaco(
  data_dir,
  slice,
  filename,
  variable_features_n = variable_features_n,
  spatial_file = spatial_file,
  vars_to_regress = NULL
)
}
\arguments{
\item{data_dir}{Directory containing the H5 file specified by file name and the image data in a sub directory called spatial}

\item{slice}{Name for the stored image of the tissue slice}

\item{filename}{Filename of data to be read}

\item{variable_features_n}{Number of most variable features to keep.}

\item{spatial_file}{Name of csv file from which to read tissue positions.}

\item{vars_to_regress}{Names of features to be regressed against using \link[Seurat]{PercentageFeatureSet}}
}
\value{
Retuns a ready to run SPaCoObject.
}
\description{
Read in 10x Visium spatial transcriptomics data
}
