% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSCA.R
\name{RunMultiSCA}
\alias{RunMultiSCA}
\title{Multi slide SCA_function}
\usage{
RunMultiSCA(
  SpaCoObject,
  PC_criterion = "percent",
  PC_value = 0.9,
  compute_nSpacs = FALSE,
  nSim = 1000,
  nSpacQuantile = 0.05
)
}
\arguments{
\item{PC_criterion}{criterion on which to select number of principal components for initial covariance matrix reconstruction; either "number" to select a number of PCs or "percent" to select number of PCs to explain specified amount of data variance}

\item{PC_value}{Value to specify number of PCs or desired level of explained variance, see "PC_criterion"}

\item{compute_nSpacs}{Boolean if number of relevant spacs is to be computed. Increases run time significantly}

\item{nSim}{Number of simulations for computation of spac number}

\item{nSpacQuantile}{Quantile to use as cutoff for spac number}

\item{data}{gene expression data matrix; p genes as columns, n loci as rows}

\item{neighbourindexmatrix}{n x n matrix showing neighborhood weight of loci}

\item{compute_projections}{Boolean if meta genen projections should be computed. May increase run time significantly. Default is TRUE}
}
\value{
Returns a SpaCoObject filled with the result of the spatial component analysis.
}
\description{
Multi slide SCA_function
}
