% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA_function}
\alias{SCA_function}
\title{Title}
\usage{
SCA_function(
  data,
  neighbourindexmatrix,
  PC_criterion = "number",
  PC_value = 50,
  testStatType = "C",
  SC_criterion = "percent",
  SC_value = 0.8,
  orthogonalResult = FALSE
)
}
\arguments{
\item{data}{some text}

\item{PC_criterion}{some text}

\item{PC_value}{some text}

\item{testStatType}{some text}

\item{SC_criterion}{some text}

\item{SC_value}{some text}

\item{orthogonalResult}{some text}

\item{neighbourindexmatri}{some text}
}
\value{
some return
}
\description{
Title
}
\examples{
some example
}
