% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_projection.R
\name{set_projection}
\alias{set_projection}
\title{Set projections of Spaco Object}
\usage{
set_projection(SpaCoObject, data, neighborhood = SpaCoObject@neighbours)
}
\arguments{
\item{SpaCoObject}{Fitted object of class \code{SpaCoObject} as returned from [SpaCo] on which to update the projection slot with new data.}

\item{data}{gene expression data matrix; p genes as columns, n loci as rows}

\item{neighborhood}{neighborhood matrix of the new data. Must be of dimension n times n, where n is the number of columns of \code{data}}
}
\value{
Returns the SpaCoObject with the updated data, neighborhood matrix, and projection matrix.
}
\description{
Set projections of Spaco Object
}
