% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{smooth_profiles}
\alias{smooth_profiles}
\title{computed smoothed gene profiles of genes present in the data.}
\usage{
smooth_profiles(SpaCoObject)
}
\arguments{
\item{SpaCoObject}{Spaco object to compute profiles of.}
}
\value{
smoothed gene profiles in the SpaCoObject.
}
\description{
computed smoothed gene profiles of genes present in the data.
}
