% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{smoothed_projection_plot}
\alias{smoothed_projection_plot}
\title{Plot smoothed gene expression}
\usage{
smoothed_projection_plot(
  SpaCoObject,
  features = NULL,
  ncol = NULL,
  combine = TRUE
)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{spac}{component to plot}
}
\value{
returns a ggplot object with the smoothed gene expression.
}
\description{
Plot smoothed gene expression
}
