% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCo.R
\name{SpaCo}
\alias{SpaCo}
\title{SpaCo}
\usage{
SpaCo(
  SpaCoObject,
  PC_criterion = "percent",
  PC_value = 0.95,
  set_nspacs = NULL,
  nSim = 1000,
  nSpacQuantile = 0.05,
  reducedSpots = FALSE,
  nReduce = 1000
)
}
\arguments{
\item{SpaCoObject}{Object of class \code{SpaCoObject} as generated from [SpaCoObject] on which to perform SCA.}

\item{PC_criterion}{criterion on which to select number of principal components for initial covariance matrix reconstruction; either "number" to select a number of PCs or "percent" to select number of PCs to explain specified amount of data variance}

\item{PC_value}{Value to specify number of PCs or desired level of explained variance, see "PC_criterion"}

\item{set_nspacs}{Boolean if number of relevant spacs is to be computed. Increases run time significantly}

\item{nSim}{Number of simulations for computation of spac number}

\item{nSpacQuantile}{Quantile to use as cutoff for spac number}

\item{reducedSpots}{Should algorithm be run on a subset of spots? Default = FALSE}

\item{nReduce}{Number of spots to sample to run algorithm on subset if \code{reducedSpots==TRUE}. Ignored if \code{reducedSpots==FALSE}}
}
\value{
Returns a SpaCoObject filled with the result of the spatial component analysis.
}
\description{
SpaCo
}
