% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Spaco_plot}
\alias{Spaco_plot}
\title{Plot SPaCo meta genes.}
\usage{
Spaco_plot(SpaCoObject, spac = 1)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{spac}{component to plot}
}
\value{
returns a ggplot object with the meta gene expression.
}
\description{
Plot SPaCo meta genes.
}
