% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seuratwrappers.R
\name{spacs_to_seurat}
\alias{spacs_to_seurat}
\title{transfer computed spatial components to existing Seurat object.}
\usage{
spacs_to_seurat(SpaCoObject, SeuratObject, nSpacs = SpaCoObject@nSpacs)
}
\arguments{
\item{SpaCoObject}{SpaCoObject to export spatial components from.}

\item{SeuratObject}{Seurat object to add spatial components to.}

\item{nSpacs}{Number of Spacs which are to be projected on for dimension reduction}
}
\value{
Returns a Seurat Object with the spatial components projections in the dimensional reduction slot.
}
\description{
transfer computed spatial components to existing Seurat object.
}
