% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalFunctions.R
\name{SVGTest}
\alias{SVGTest}
\title{Compute the spatial variable genes of a SpaCoObject after runinng runSCA}
\usage{
SVGTest(SpaCoObject, adjustMethod = "holm")
}
\arguments{
\item{SpaCoObject}{SpaCoObject to compute spatially variable genes of.}

\item{adjustMethod}{method for p-value adjustment. See p.adjust function.}
}
\value{
returns a data frame of spatial variable genes and their p-Values.
}
\description{
Compute the spatial variable genes of a SpaCoObject after runinng runSCA
}
