% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{train_prepare}
\alias{train_prepare}
\title{Prepare Input Data for Model Training}
\usage{
train_prepare(
  X,
  OBS,
  T1,
  Z,
  K,
  run_prepare = TRUE,
  run_initial = TRUE,
  random_number = 10,
  random_state = 0,
  mean_removal = TRUE,
  homo_noise = TRUE,
  fit_intercept = TRUE,
  initial_type = "pca",
  update_sigma_noise = TRUE,
  update_sigma_mu = TRUE,
  update_smooth_penalty = TRUE,
  update_lasso_penalty = TRUE,
  lambda1_dfmin = NULL,
  lambda1_dfmax = NULL,
  Omega = NULL,
  h = NULL,
  nlambda1 = 10,
  nlambda2 = 100,
  lambda1 = NULL,
  lambda2 = NULL,
  nlambda0 = 10,
  foldid = NULL,
  beta_folds = 10,
  kappa = 0.01,
  orthogonal = 0,
  ridge_traj = 1e-04,
  lam1_criterion = 1,
  lasso_max_iter = 10000,
  lasso_tol = 1e-08,
  eps = 1e-10
)
}
\arguments{
\item{X}{An \code{I} (number of subjects) × \code{T} (number of times) × \code{J} (number
of features) array containing time-series data.}

\item{OBS}{An \code{I} (number of subjects) × \code{T} (number of times) × \code{J} (number
of features) array indicating whether an observation is available in array
\code{X}.}

\item{T1}{A length \code{T} (number of times) vector containing measured time.}

\item{Z}{An \code{I} (number of subjects) × \code{q} (number of covariates) matrix
containing auxiliary covariates.}

\item{K}{Selected rank.}

\item{run_prepare}{Whether to prepapre data. Default is TRUE.}

\item{run_initial}{Whether to perform parameter initialization. Default is
TRUE.}

\item{random_number}{Random integer for random factor matrix initialization
in PARAFAC decomposition. A positive value indicating running random
initialization and comparing the errors with svd initialization. Default is
10.}

\item{random_state}{Control the randomness of random parameter
initialization. Default is 0.}

\item{mean_removal}{Whether to remove the mean time curve. Default is TRUE.}

\item{homo_noise}{Whether to include the homogeneous noise variances for
\code{Xj}. Default is TRUE.}

\item{fit_intercept}{Whether to calculate the intercept for the regression
model with lasso penalty when updating parameter \code{beta}. Default is TRUE.}

\item{initial_type}{Type of parameter initialization. Default is "pca".}

\item{update_sigma_noise}{Whether to update the variances for \code{Xj}. Default
is TRUE.}

\item{update_sigma_mu}{Whether to update the variances for \code{mu}. Default
is TRUE.}

\item{update_smooth_penalty}{Whether to update the smooth penalty. Default
is TRUE.}

\item{update_lasso_penalty}{Whether to update the lasso penalty. Default
is TRUE.}

\item{lambda1_dfmin}{Minimum degrees of freedom for modeling the
trajectories.}

\item{lambda1_dfmax}{Maximum degrees of freedom for modeling the
trajectories.}

\item{Omega}{Smooth penalty matrix.}

\item{h}{Norm of \code{Phi}.}

\item{nlambda1}{Number of \code{lambda1} values. Default is 10.}

\item{nlambda2}{Number of \code{lambda2} values. Default is 100.}

\item{lambda1}{A length \code{K} vector for smoothness penalty.}

\item{lambda2}{A length \code{K} vector for lasso penalty.}

\item{nlambda0}{Number of \code{lambda} values in data preparation. Default is 10.}

\item{foldid}{Optional vector of values between 1 and \code{beta_folds}
identifying what fold each observation is in for cross-validation when
updating parameter \code{beta}.}

\item{beta_folds}{Number of folds for cross-validation when updating
parameter \code{beta}. Default is 10.}

\item{kappa}{Small identity for numerical stability when encouraging the
smoothness penalty. Default is 1e-2.}

\item{orthogonal}{Constraints on the orthogonality between \code{Phi}s. Default
is 0.}

\item{ridge_traj}{Ridge trajectory when updating parameter \code{Phi}. Default
is 1e-4.}

\item{lam1_criterion}{Extra standard deviation away from the minimum cross
validation error. Default is 1.}

\item{lasso_max_iter}{Maximum number of loop iterations when updating \code{beta}.
Default is 1e4.}

\item{lasso_tol}{Tolerance for loop iterations when updating \code{beta}. Default
is 1e-8.}

\item{eps}{Small regularization parameter to avoid severe over-fitting to
the noise. Default is 1e-10.}
}
\value{
A list containing required elements for model training.
}
\description{
This function prepares user-level input data for model training.
}
\examples{
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
spaco_object <- train_prepare(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                              Z = impact$Z, K = 4, mean_removal = FALSE)
}
