% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{train_spaco}
\alias{train_spaco}
\title{SPACO Model Training}
\usage{
train_spaco(
  spaco_object,
  max_iter = 100,
  min_iter = 1,
  tol = 1e-06,
  update_sigma_noise = TRUE,
  update_sigma_mu = TRUE,
  update_smooth_penalty = TRUE,
  update_lasso_penalty = TRUE,
  trace = FALSE,
  reorder = TRUE
)
}
\arguments{
\item{spaco_object}{A list containing required elements for model training.
This is the output of \code{train_prepare} function.}

\item{max_iter}{Maximum number of loop iterations in parameter updates.
Default is 100.}

\item{min_iter}{Minimum number of loop iterations in parameter updates.
Default is 1.}

\item{tol}{Tolerance for loop iterations in parameter updates. Default is
1e-6.}

\item{update_sigma_noise}{Whether to update the variances for \code{Xj}. Default
is TRUE.}

\item{update_sigma_mu}{Whether to update the variances for \code{mu}. Default
is TRUE.}

\item{update_smooth_penalty}{Whether to update the smooth penalty. Default
is TRUE.}

\item{update_lasso_penalty}{Whether to update the lasso penalty. Default
is TRUE.}

\item{trace}{Whether to print the model training process represented by the
sum of errors in parameter updates. Default is TRUE.}

\item{reorder}{Whether to reorder the parameters. Default is TRUE.}
}
\value{
A list containing the results of model training.
}
\description{
This function trains SPACO model on the prepared input data.
}
\examples{
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
spaco_object <- train_prepare(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                              Z = impact$Z, K = 4, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
}
