% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seuratwrappers.R
\name{create_SpaCoObject_from_KNN}
\alias{create_SpaCoObject_from_KNN}
\title{runSCA_for_uMAP}
\usage{
create_SpaCoObject_from_KNN(Seurat)
}
\arguments{
\item{SeuratObject}{}
}
\description{
runSCA_for_uMAP
}
