% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{denoised_projection_plot}
\alias{denoised_projection_plot}
\title{Plot denoised gene expression}
\usage{
denoised_projection_plot(
  SpaCoObject,
  features = NULL,
  ncol = NULL,
  combine = TRUE
)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{features}{feature to plot}

\item{ncol}{Number of columns for combined plot if \code{combine = TRUE}.}

\item{combine}{Should spac plots be combined? Default = TRUE}
}
\value{
returns a ggplot object with the denoised gene expression.
}
\description{
Plot denoised gene expression
}
