% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{denoised_projection_plot}
\alias{denoised_projection_plot}
\title{Plot denoised gene expression}
\usage{
denoised_projection_plot(
  SpaCoObject,
  features = NULL,
  ncol = NULL,
  combine = TRUE
)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{spac}{component to plot}
}
\value{
returns a ggplot object with the denoised gene expression.
}
\description{
Plot denoised gene expression
}
