% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{feature_plot}
\alias{feature_plot}
\title{Plot gene expression}
\usage{
feature_plot(SpaCoObject, features = NULL, ncol = NULL, combine = TRUE)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{features}{gene to plot}
}
\value{
returns a ggplot object with gene expression.
}
\description{
Plot gene expression
}
