% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{feature_plot}
\alias{feature_plot}
\title{Plot for IMPACT Dataset}
\usage{
feature_plot(spaco_object, selected_feature, columns_feature)
}
\arguments{
\item{spaco_object}{A list containing the results of model training. It is
assumed to include elements \code{num_subjects}, \code{num_times}, \code{X}, \code{OBS}, and
\code{Xhat}.}

\item{selected_feature}{Name of the selected feature.}

\item{columns_feature}{Name of all features.}
}
\value{
A plot of the observed versus estimated values.
}
\description{
This function plots the observed versus estimated values for
the selected feature.
}
\examples{
# IMPACT
data("impact_imputed")
data("impact_missing")
impact <- impact_data_wrangling(impact_missing, impact_imputed)
spaco_object <- train_prepare(X = impact$X, OBS = impact$OBS, T1 = impact$T1,
                              Z = impact$Z, K = 4, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
spaco_object <- feature_predict(spaco_object)
feature_plot(spaco_object, "TcellsofLivecells", impact$columns_feature)
feature_plot(spaco_object, "TotalNeutrophilsofLivecells",
            impact$columns_feature)
feature_plot(spaco_object, "HLA.DR.ofTotalMono", impact$columns_feature)
feature_plot(spaco_object, "IL6", impact$columns_feature)

# IMMUNE
data("immune_original")
immune <- immune_data_wrangling(immune_original)
spaco_object <- train_prepare(X = immune$X, OBS = immune$OBS, T1 = immune$T1,
                              Z = immune$Z, K = 5, mean_removal = FALSE)
spaco_object <- train_spaco(spaco_object, max_iter = 30, min_iter = 1,
                            tol = 1e-4, trace = TRUE)
spaco_object <- feature_predict(spaco_object)
feature_plot(spaco_object, "IgG..LY", immune$columns_feature)
feature_plot(spaco_object, "CD19._CD20..LY.1", immune$columns_feature)
feature_plot(spaco_object, "CD4._CD8..CD3.", immune$columns_feature)
feature_plot(spaco_object, "LY.All_CD45", immune$columns_feature)
}
