% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_for_spaco.R
\name{read_10x_for_spaco}
\alias{read_10x_for_spaco}
\title{Read in 10x Visium spatial transcriptomics data}
\usage{
read_10x_for_spaco(
  data_dir,
  slice,
  filename,
  only_var = TRUE,
  variable_features_n = variable_features_n
)
}
\arguments{
\item{data_dir}{Directory containing the H5 file specified by file name and the image data in a sub directory called spatial}

\item{slice}{Name for the stored image of the tissue slice}

\item{only_var}{Logical True if you want to keep only the most variable features in the SPACO object. Default True.}

\item{variable_features_n}{Number of most variable features to keep.}
}
\description{
Read in 10x Visium spatial transcriptomics data
}
