% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA_function}
\alias{SCA_function}
\title{SCA_function}
\usage{
SCA_function(
  SpaCoObject,
  PC_criterion = "percent",
  PC_value = 0.8,
  orthogonalResult = FALSE,
  compute_projections = F
)
}
\arguments{
\item{PC_criterion}{criterion on which to select number of principal components for initial covariance matrix reconstruction; either "number" to select a number of PCs or "percent" to select number of PCs to explain specified amount of data variance}

\item{PC_value}{Value to specify number of PCs or desired level of explained variance, see "PC_criterion"}

\item{orthogonalResult}{Logical value to specify if Spacos should be orthogonalized to form a ONB; since transformation of eigenvalues results in non-orthogonal Spacos}

\item{data}{gene expression data matrix; p genes as columns, n loci as rows}

\item{neighbourindexmatrix}{n x n matrix showing neighborhood weight of loci}
}
\description{
SCA_function
}
