% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{smoothed_projection_plot}
\alias{smoothed_projection_plot}
\title{Title}
\usage{
smoothed_projection_plot(
  SpaCoObject,
  features = NULL,
  ncol = NULL,
  combine = TRUE
)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{spac}{Spatial component to plot}
}
\description{
Title
}
