% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{Spaco_plot}
\alias{Spaco_plot}
\title{Title}
\usage{
Spaco_plot(SpaCoObject, spac = 1, ncol = NULL, combine = TRUE)
}
\arguments{
\item{SpaCoObject}{SpacoObject with computed projections}

\item{spac}{component to plot}
}
\description{
Title
}
